split cancel popup into another file
